Rename module name to plural for semantic consistency
